+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.1.0]
+
+- New Features:
+  - Buffer Array disposing utilities in GraphicsUtil
+  - common stride sizes in a static class called StrideSize
+  - ComputeShaderUtil for handling simple parallel task
+
 ## [1.0.0]
 
 - Initial release